Return observation on `reset` executed Add observation_space, action_space
